esgdsd_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
asyncpg==0.29.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytz==2023.3
email_validator==2.1.0
requests==2.31.0
httpx==0.25.1
pdf2image==1.16.3
PyMuPDF==1.23.7
pytesseract==0.3.10
pdfminer.six==20221105
Pillow==10.1.0
opencv-python==4.8.1.78
torch==2.1.1
transformers==4.35.2
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
openpyxl==3.1.2
openai==1.3.0 